# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Social Media API",
    "version": "1.0.0",
    "description": "Professional REST API documentation for the Social Media Backend. Supports authentication, posts, comments, stories, notifications, real-time messaging, and more.",
    "contact": {
      "name": "API Support",
      "email": "support@socialmedia.dev"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current Environment API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "أدخل رمز JWT: Bearer <token>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fail",
              "error"
            ],
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          }
        }
      },
...
      "UserPublic": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65a1b2c3d4e5f6a7b8c9d0e1"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "displayName": {
            "type": "string",
            "example": "John Doe"
          },
          "avatar": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri",
                "example": "https://res.cloudinary.com/..."
              }
            }
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "bio": {
            "type": "string",
            "example": "Software engineer & coffee enthusiast"
          },
          "followersCount": {
            "type": "integer",
            "example": 1240
          },
          "followingCount": {
            "type": "integer",
            "example": 380
          }
        }
      },
...
      "PostMedia": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "image",
              "video"
            ],
            "example": "image"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "thumbnail": {
            "type": "string",
            "format": "uri"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "width": {
                "type": "integer"
              },
              "height": {
                "type": "integer"
              },
              "size": {
                "type": "integer"
              },
              "format": {
                "type": "string"
              }
            }
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "65a1b2c3d4e5f6a7b8c9d0e1"
          },
          "author": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "content": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string",
                "example": "Hello world! #social @friend"
              }
            }
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMedia"
            }
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "friends",
              "followers",
              "private",
              "custom"
            ],
            "example": "public"
          },
          "likesCount": {
            "type": "integer",
            "example": 42
          },
          "commentsCount": {
            "type": "integer",
            "example": 7
          },
          "sharesCount": {
            "type": "integer",
            "example": 3
          },
          "isLiked": {
            "type": "boolean",
            "example": false
          },
          "hashtags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "social",
              "tech"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isEdited": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Please provide a valid email address"
                },
                "value": {
                  "type": "string",
                  "example": "not-an-email"
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "fail",
              "message": "Invalid token. Please log in again."
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "fail",
              "message": "Post not found"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden — insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request — validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Something went wrong!"
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/posts/{postId}/comments": {
      "post": {
        "summary": "Add a text comment to a post",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "parentCommentId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added"
          }
        }
      },
      "get": {
        "summary": "Get comments for a post",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments"
          }
        }
      }
    },
...
    "/posts": {
      "post": {
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "description": "Post text content"
                  },
                  "media": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Up to 10 media files (images or videos)"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "friends",
                      "followers",
                      "private",
                      "custom"
                    ],
                    "default": "public"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "post": {
                          "$ref": "#/components/schemas/Post"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "summary": "Get a specific post by ID",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB ObjectId of the post"
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "post": {
                          "$ref": "#/components/schemas/Post"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update a post (author only)",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "maxLength": 5000
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "friends",
                      "followers",
                      "private",
                      "custom"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "post": {
                          "$ref": "#/components/schemas/Post"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete a post (author or admin)",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Post deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/posts/{postId}/like": {
      "post": {
        "summary": "Like or unlike a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated like status"
          }
        }
      }
    },
    "/posts/{postId}/share": {
      "post": {
        "summary": "Share a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Post shared"
          }
        }
      }
    },
    "/posts/{postId}/viral-stats": {
      "get": {
        "summary": "Get viral coefficient (K-factor) and reach stats for a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Viral stats including K-factor, total reach, share depth"
          }
        }
      }
    },
    "/posts/{postId}/share-tree": {
      "get": {
        "summary": "Get the share propagation tree for a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Share tree with reshares and their engagement"
          }
        }
      }
    },
    "/posts/capsules/mine": {
      "get": {
        "summary": "List my sealed time capsule posts",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending capsule posts with reveal dates"
          }
        }
      }
    },
    "/posts/{postId}/seal": {
      "post": {
        "summary": "Seal a post as a time capsule",
        "description": "The post will be hidden until the revealAt date, then automatically published.",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "revealAt"
                ],
                "properties": {
                  "revealAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2027-01-01T00:00:00Z"
                  },
                  "hint": {
                    "type": "string",
                    "maxLength": 150,
                    "example": "Something big is coming..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post sealed successfully"
          }
        }
      },
      "delete": {
        "summary": "Cancel (unseal) a time capsule before reveal",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Capsule cancelled and post restored"
          }
        }
      }
    },
...
  }
}